Run cleanup pass cleanUpTransModes after every pass were it could be needed.
Things are simpler this way.  It will also make it easier to explain
the algo as cleanUpTransModes no longer needs to be explained as a
separate pass.

Also of note that this change is also a workaround for a bug
somewhere that should probably be looked into at some point.

Towards ACUMEN-257.