Added function that can be used to draw attention to UI components.
The function "flashes" a color passed to its function argument, by
varying the passed color according to a pulse. This can e.g. be
used to flash the background color of a text field when the input
is invalid.

Used function to flash the background of the search text field when
the search term is not found in the model, and the console tab when
an error is reported in the console (useful when the console tab is
not selected).