Updated regression tests w.r.t. semantics change.
The change in the way that continuous assignments (ODEs)
are desugared changes the way some models simulate.
In this commit, the regression test reference data files
have been updated for eight models that have been verified
to simulate correctly with the new semantics. The tests of
three models have also been disabled, given that the models
need to be updated w.r.t. the semantics change.

 Updated regression data file:

  1. 0_Demos/00_Welcome_to_Acumen.acm
  2. 0_Demos/03_Car_and_bike.acm
  3. 0_Demos/04_Car_followed_by_camera.acm
  4. 0_Demos/05_RC_circuit.acm
  5. 2_Modeling_Physical_Systems/iccps_mass_pendulum.acm
  6. 7_Single_Link_Robot/01_Converting_Accelerations.acm
  7. 9_Experimental/hodgkin-huxley.acm
  8. A_Ping_Pong/2012_hh_1M.acm

 Disabled test pending update to model:

  1. A_Ping_Pong/2012_hh_2M.acm
  2. A_Ping_Pong/2012_hh_3M.acm
  3. A_Ping_Pong/2012_hh_4M.acm

This relates to ACUMEN-314.