Check for duplicate cont. assignments in standard reference interpreter.
This was implemented by using the same extension of the Eval type
used in interpreters.reference.experimental.Interpreter. This makes it
possible to share the Eval type between the two interpreters.
In the standard interpreter, the 'eqs' part of the Eval tuple is simply
used to book-keep continuous assignments.

This fixes ACUMEN-321.