Rethink how the PI constant is handled so its less of a special case.
That is, PI is now a GDouble rather than a separate GroundValue type.
The only special case now is in the enclosure interpreter, which tests
for the special value (acumen.CONST_PI) and uses its own representation.

This fixes ACUMEN-328 so that -pi now works.  Since PI is no longer a
special case this also allows pi to correctly print in the pretty
printer (and thus one example result needed to be updated).