Resolve modes using preconds & intelligently combine preconds when merging.
The target mode of a reset is now resolved based on matching the
postconds of the reset with the preconds of the the mode.  In order
for this to work the special mode variable is now part of the preconds
and the mode's preconds are intelligently combined when merging one
node into another.

Also, in mergeDupModes use resets, not preconds, as the basis for
determining if two modes can be combined (similar to what is is now
being done in cleanUpTransModes).

Towards ACUMEN-349.