Corrected duplicate discrete assingment check.
Now all discrete assignments are registered (even identify
assignments, i.e. those that do not change the Store), in order
to always detect duplicate assignments to the same lhs.

Also, duplicate assingments are now reported using a custom
AcumenError, i.e. an object of a subclass to DuplicateAssingment.

This change was implemented in both the standard and experimental
reference interpreters.

Test cases added based on model attached to ACUMEN-348.