Add option to enable extract-ha and related passes via U.I.
The new menu option is called "Normalization".

Also introduce the idea of aliases in the pass management so that one
pass can be an alias for another pass or a set of passes.  This is
then used to easily enable the set of passes associated with the
"Normalization" menu option.  In addition, the "normalization" passes
can be selected from the command line without any additional code.

Towards ACUMEN-257 and ACUMEN-355.