Separate ODE solving from AST eval. in Traditional (Exp) semantics.
Now, ODE solving is delayed by saving all information necessary to
solve the IVP (e.g. the environment in which the RHS is evaluated)
and performing the solving at the end of the integration step.

This is done in preparation for implementing additional ODE
solvers for this semantics, configurable through a simulator
parameter.

This relates to ACUMEN-364.