Fix bug that was causing the Runge Kutta integrator to effectively implement Forward Euler.
Because of the way in which higher-order ODEs were being desugared into a system of
first-order ODEs (more specifically, that the field of the second-highest-order variable
was set to the highest-order variable), the ODE that was being passed to the solvers
had a constant field.

This was fixed by switching desugaring of ODEs off in Desugarer.scala and performing
the corresponding expansion in the interpreter. Now, instead of in-lining the value of the
highest-order variable, the RHS expression of the highest-order variable is passed to
the ODE solver as the field of the second-highest-order variable. The result is a correctly
functioning Runge Kutta ODE solver.

A side effect of this change is that the Eval monad now no longer separately stores ODEs,
as the corresponding EquationI's are no longer emitted by Desugarer.scala. ODEs are
instead filtered out from other continuous assignments by looking at the number of primes
in their Names.

This relates to ACUMEN-374.