Wrap program generator in function that attempts to fix invalid models.
The generator genProg now calls ensureGoodProg, which in turn calls
simulationThrowsError on the generated program p, to check whether
simulation of p using the reference interpreter produces any errors.
If it does, ensureGoodProg will attempt to modify p in order to make
it into a valid program. Currently, this is limited to one kind of error:
if a ContinuousDynamicsUndefined is encountered, ensureGoodProg
will add a continuous assignment to the corresponding variable to p
and then call itself with the updated model.

This relates to ACUMEN-370.