Re-implement in-lining of ODEs into EquationI using de-sugarer.
Use a modified version of the original desugaring approach for ODES,
where EquationIs are emitted locally, right where the ODE is located in
the AST. Now, instead of emitting the LHS of the ODE as the RHS of
the highest-order EquationI, emit the RHS of the ODE directly.

This relates to ACUMEN-370.