Revert to evaluating continuous assignments during AST traversal.
Evaluate the RHS of a continuous assignment and log the resulting
value in Eval monad directly the when it is processed during AST
traversal. This is instead of delaying the evaluation by logging the
RHS with its environment. This change is a roll-back to the way this
worked before commit. d6768bf.

This relates to ACUMEN-370.