Add support for "include" directive.
The syntax is:
  include "some_file.acm";
and all includes must be at the start of the file for now.

File names in the include are relative to the current location of the
source file.

Includes are only supported when the enhanced parse method in SemanticsImpl
is used.  The default entry point to the parser does not support
processing includes.

In order for error messages to still make since, the position information
stored in the source code and error messages was enhanced to include
the source file name.  In order to implement this it was necessary to
reimplement StreamReader as that is a sealed class in the API.

Towards ACUMEN-237.