Update the optimized interpreter to implement the Reference 2014 semantics using the IVPRungeKutta ODE solver.
This was done by changing the parCont parameter to contMode that can
now be one of ContMode.Seq, Par or IVP.  When the new IVP parameter is
used the EquationI are collected into a vector during the conscious
phase.  After collecting the equations the IVPRungeKutta solver is ran
with the collected vector serving as the Field.  The RichStore used by
the solver is represented as a vector of values.

The implementation of the operations on Field and RichStore is such
that all computation is performed eagerly rather than lazily as is
done by the Reference 2014 interrupter.

Also: Updated SemanticsImpl to support the new "optimized2014"
semantics; added the corresponding item to the GUI menu; and added
some test case for the new semantics.