Factor out semantics tests from interpreter tests.
Factor out interpreter tests that are really tests of the
implementation of a semantics and not specifically the interpreter into
a separate test.  The test is named based on the internal name of the
interrupter/semantics with a suffix of "SemanticsTest", for example the
semantics test for the Reference 2014 interpreter is
"Reference2014SemanticsTest".  The tests are stored in the same
package of the interpreter being tested.

Common code for the semantics tests is in the "semanticstest" package.

The idea behind this refactoring is to provide a consistent set of
tests for an interpreter that implements a given semantics.  For
example the Reference 2014 and Optimized interpreter both implement
the same semantics and hence should be subject to the same tests.