Implement basic #semantics directive.
The string currently used is the internal name for the semantics.  For
example:
  #semantics "reference2014"
Will select the "2014 Reference" semantics.

For now, the semantics directive takes precedence over the semantics
chosen via the menu or on the command line.

To avoid confusion the currently used semantics is outputted to the
console when playing a model.

Towards ACUMEN-38.