Redo how Semantics are selected from a string.
Each semantics now has an internal string, a display string, and any
number of alias.  Any one of them can be used in for the semantics
string specified with '#semantics' directive or the --semantics
command line option.

The display name is now used when reporting the semantics in the console.

Also change the error message from UnrecognizedInterpreterString to
UnrecognizedSemanticsString and make it positional so that line and
column information can be reported.

Towards ACUMEN-38.