Purely functional implementation of direct enclosure semantics.
This implementation is not based on the Eval monad used in the
Reference 2014 interpreter. Instead, stores are explicitly threaded
through the interpreter.
Information about active statements, that was previously stored
in the Eval type, is now stored in the type Changeset, and the
AST traversal (evalAction etc.) yields a Set[Changeset].
Instead of updating the store during traversal, the interpreter's
step function creates an updated store based on the one produced by
the previous step previous and the Changesets that were collected
during AST traversal.

This relates to ACUMEN-415.