Delay evaluation of action expressions.
Previously, DelayedActions would store a pre-computed value
that was written to a store (Enclosure) using EnclosureOps.apply().
Now, DelayedAction stores the RHS of assignments (and ODEs) as well
as the Env that is necessary to evaluate it. The evaluation of the
RHS is done in EnclosureOps.apply(). Not storing the value of the
RHS in the DelayedAction makes it possible to safely compare
values of the DelayedAction type using structural equality (the
equality used to compate case classes), making the function
sameChange() unnecessary.

This relates to ACUMEN-415.