Tag initial conditions with Evolution instead of Changeset.
This makes it possible to group, and hence merge, initial
conditions according to the evolution that yielded them.
This merge criterion allows the "no event" branch to catch up with
the "event" branch, as long as there is a sufficient number of flow
time steps before the next event happens.

This relates to ACUMEN-415.