Prompt the user when the specified semantics does not match the selected one.
When the semantics specified in the model does not match what is
currently selected in the U.I., prompt the user.  If the user clicks
"OK" than change the Menu to match the model.  Otherwise abort to
allow the user to change what is specified in the model.

Towards ACUMEN-38.