Initial implementation of Generic logger Interface.
Rewrote the Console interface to make it simpler and to use the Logger
data structures.  The proper way to write to the console is through
the logger.

Warning, and debug messages are not really supported yet.  If one
of these messages is sent the console will treat it as a normal message.
The backTrace field in Message is also ignored for now.

See ACUMEN-233.