Bug fix in optimized enclosure plotting code when plotting thin enclosures.
For example, in this model:

  class Main(simulator)
    private
      x := 0; x' := 1;
      c := 5
    end
    x' = 1
  end

The plots for x' and c variables are blank.  This was because the
enclosures where thin (ie an exact solution) and had no area.
ava.awt.geom.Area will not show anything in this case.  The solution
was to instead plot the outline directly.

Towards ACUMEN-421.