Add contraction of GStrEnclosures.
Based on how contraction is done in acumen.interpreters.enclosure.Contract, for
the "==" and "~=" operators. Note that all that is done for "~=" is to check if the
result is an empty box, which useful for performing intersections, but can not be
used as a constraint to improve an enclosure.

This relates to ACUMEN-453.