Simplify the release script and add support for a "build id".
The simplification is the removal for support to pushing to the
special effective/acumen repo. that is used to hold the censored
version of the source.

In addition a unique build-id is created for each build.  The build-id
is the number of seconds from 2010-01-01 00:00 UTC combined with the
first eight digits of the commit id.  This file is written to
src/main/resources/acumen/build_id in a similar way the version is
written to src/main/resources/acumen/version.

Towards ACUMEN-486.