Delay creates and associated reparentings.
Creates are now applied at the end of a step, similar to how assignments
are processed. Because creates are delayed, it is no longer possible to
compute the list of orphans (children of a dead object, that need to be
reparented to become children of the dead object's parent) during the
AST traversal. This made it necessary to change the order in which
terminations and reparenting are processed during the Discrete phase.
Now, terminations are applied first, yielding the list of orphans, which
are then processed together with the remaining reparentings.