Move add, sub and neg into concrete Dif types.
To make the implementation more homogeneous.

Also, remove coeff as a member of Dif[V], to avoid accidental
direct uses of this collection, that would not obey the contract
of Dif[V].apply(i), that should return the zero of V for any i
greater than Dif[V].length().

This relates to ACUMEN-620.