introduce web interface
This commit introduces the web interface. It uses sockets to
connect to a node.js client, which in turn connects to a javascript
client via websockets.

All changes included in this commit:
- Modified README.md and add new gui image
- Created web interface's files
- Migrated progress bar functionality
- Migrated the table tab
- Migrated the file browser tab
- Migrated editor functionality
- Migrated most menu items

Migrated items were removed from the original acumen's swing gui
to prepare for the complete deprecation of the library.