implement 3D data exporting to web interface
Acumen now creates and exports the 3D rendering data using the json
format, which is passed around as a javascript object. Implementing
a 3D rendering library should now be able to read it and render
the complete animation. As of right now, this json format
contains the whole frameset and is constructed and sent after
the original animation is over.

Changes made with this commit:
- Exporting 3D data to json formatted file
- Import the 3D data from the web interface (javascript)
- Automatic selection of _3D tab when animation is identified
- Implemented dynamic plot and 3D tab selection status

There are also TODO notes that point to the creation of the 3D
dataset model which represents the complete animation. Making
use of this model directly should eliminate the need to wait
for the animation to finish before sending the json object.