handle errors concerning indexing options
This commit aims to resolve and better inform the user of the type of
error while using various indexing options. Instead of a simple number
which describes the index position the error was found, the console now
correctly informs about both the position and the variable which can
not be evaluated. In detail:

1) On index out of bounds:
  - Instead of just the index position, the console now displays the
    variable as well as the index position.
  - If vector is n-dimensional, properly display the entire index
    position. For e.g: 'Index (0,1) is out of bounds'.
2) Trying to index into a construct which is not any kind of list:
  - The console now properly displays a message that only indexing on
    vectors is allowed.